Feat. Validates the parameters

+++ controller.php
@@ -8,8 +8,7 @@
 require_once('TodoList.php');
 require_once('utils.php');
 
-class controller
-{
+class controller {
     protected $ci;
 
     public function __construct($ci) {
@@ -28,6 +27,10 @@ class controller
     public function getTaskByID($request, $response, $args) {
         $id = $args['id'];
         $todoList = $this->ci->todoList;
+
+        if (!$todoList->isTaskIDExist($id))
+            return $this->IDNotFound($request, $response, $id);
+
         $data = $todoList->getTaskByID($id);
 
         $response = $response->withStatus(200); //OK
@@ -36,9 +39,11 @@ class controller
     }
 
     public function createTask($request, $response, $args) {
-        $body = $request->getParsedBody();
-        $content = $body['content'];
+        $content = $this->requireBodyContent($request, 'content');
         $todoList = $this->ci->todoList;
+
+        if ($content === false)
+            return $this->illegleArgument($request, $response, $args);
 
         $id = $todoList->createTask($content);
         $entry = $todoList->getTaskByID($id);
@@ -48,30 +53,57 @@ class controller
         return $response;
     }
 
-    public function updateTask($request, $response, $args){
-        $body = $request->getParsedBody();
-        $content = $body['content'];
+    public function updateTask($request, $response, $args) {
+        $content = $this->requireBodyContent($request, 'content');
         $todoList = $this->ci->todoList;
         $id = $args['id'];
 
-        $todoList->updateTask($id,$content);
-        $entry = $todoList->getTaskByID($id);
+        if (!$todoList->isTaskIDExist($id))
+            return $this->IDNotFound($request, $response, $id);
+        if ($content === false)
+            return $this->illegleArgument($request, $response, $args);
 
-        $response=$response->withStatus(201); //Content Created
-        $response=$response->write(jsonFormater($entry));
+        $todoList->updateTask($id, $content);
+        $data = $todoList->getAllTasks($id);
+
+        $response = $response->withStatus(201); //Content Created
+        $response = $response->write(jsonFormater($data));
         return $response;
     }
 
-    public function deleteTask($request, $response, $args){
+    public function deleteTask($request, $response, $args) {
         $id = $args['id'];
         $todoList = $this->ci->todoList;
+
+        if (!$todoList->isTaskIDExist($id))
+            return $this->IDNotFound($request, $response, $id);
+
         $todoList->deleteTask($id);
 
         $response = $response->withStatus(204); //No content
         return $response;
     }
 
-    public function noContent($response){
+    public function IDNotFound($request, $response, $args) {
+        $payload = json_encode(["error" => "No entry by this ID: " . $args]);
+
+        $response = $response->withStatus(404);//Not Found
+        $response = $response->write($payload);
         return $response;
+    }
+
+    public function illegleArgument($request, $response, $args) {
+        $payload = json_encode(["error" => "Illegle Argument"]);
+
+        $response = $response->withStatus(400); //Invalid Request
+        $response = $response->write($payload);
+        return $response;
+    }
+
+    private function requireBodyContent($request, $content) {
+        $body = $request->getParsedBody();
+        if (empty($body)) return false;
+        if (!array_key_exists($content, $body)) return false;
+        return $body[$content];
     }
 }
